Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
